back to int32

+++ minigun/minigun/base.h
@@ -4,7 +4,7 @@
 #include <cstdint>
 #include <dmlc/logging.h>
 
-typedef int64_t mg_int;
+typedef int32_t mg_int;
 
 namespace minigun {
